Checkpoint in the big merge of the controller
