--- src/controller/request/index.js
@@ -1,5 +1,0 @@
-// __Dependencies__
-const deco = require('deco');
-
-// __Module Definition__
-module.exports = deco(__dirname, ['allow', 'validation', 'conditions', 'streams']);
